backend/vercel_deployment/requirements.txt
flask==2.3.2
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
flask-cors==3.0.10
